requirements.txt
streamlit
google-generativeai
pymupdf

